Refactor: Extract `SessionDetailsScreen` components into standalone files
- Move UI components and the `TestTag` object from `SessionDetailsScreen.kt` to a dedicated `components` package for better modularity.
- Create individual files for `TopBar`, `SessionBannerImage`, `SessionLevel`, `SessionSpeakerNameAndFavouriteIcon`, `SessionTimeAndRoom`, `SessionTitleAndDescription`, and `SpeakerTwitterHandle`.
- Update `SessionDetailsScreen` to use the newly extracted components via imports.
- Improve code organization and maintainability within the session details feature.
