+++ presentation/src/main/java/com/android254/presentation/sessionDetails/view/SessionDetailsScreen.kt
@@ -72,6 +72,15 @@ import com.android254.presentation.models.SessionDetailsPresentationModel
 import com.android254.presentation.models.SessionDetailsSpeakerPresentationModel
 import com.android254.presentation.sessionDetails.SessionDetailsUiState
 import com.android254.presentation.sessionDetails.SessionDetailsViewModel
+import com.android254.presentation.sessionDetails.view.components.CustomDivider
+import com.android254.presentation.sessionDetails.view.components.SessionBannerImage
+import com.android254.presentation.sessionDetails.view.components.SessionLevel
+import com.android254.presentation.sessionDetails.view.components.SessionSpeakerNameAndFavouriteIcon
+import com.android254.presentation.sessionDetails.view.components.SessionTimeAndRoom
+import com.android254.presentation.sessionDetails.view.components.SessionTitleAndDescription
+import com.android254.presentation.sessionDetails.view.components.SpeakerTwitterHandle
+import com.android254.presentation.sessionDetails.view.components.TestTag
+import com.android254.presentation.sessionDetails.view.components.TopBar
 import com.droidconke.chai.ChaiDCKE22Theme
 import com.droidconke.chai.atoms.ChaiRed
 import com.droidconke.chai.atoms.ChaiTeal90
@@ -233,253 +242,6 @@ fun Body(
 
         Spacer(modifier = Modifier.height(140.dp))
     }
-}
-
-@Composable
-private fun SpeakerTwitterHandle(
-    speaker: SessionDetailsSpeakerPresentationModel,
-) {
-    val context = LocalContext.current
-    val intent =
-        remember {
-            Intent(
-                Intent.ACTION_VIEW,
-                Uri.parse("https://www.twitter.com/${speaker.twitterHandle}"),
-            )
-        }
-
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween,
-    ) {
-        ChaiBodyMedium(
-            bodyText = stringResource(R.string.twitter_handle_label),
-            textColor = MaterialTheme.chaiColorsPalette.textNormalColor,
-        )
-
-        COutlinedButton(
-            onClick = { context.startActivity(intent) },
-            shape = RoundedCornerShape(10.dp),
-            colors =
-                ButtonDefaults.buttonColors(
-                    MaterialTheme.chaiColorsPalette.outlinedButtonBackgroundColor,
-                ),
-        ) {
-            Icon(
-                painter = painterResource(id = R.drawable.ic_twitter_logo),
-                contentDescription = null,
-                modifier =
-                    Modifier
-                        .height(20.dp)
-                        .width(20.dp),
-                tint = MaterialTheme.chaiColorsPalette.secondaryButtonColor,
-            )
-
-            Spacer(modifier = Modifier.width(5.dp))
-
-            ChaiBodyMedium(
-                modifier = Modifier.testTag(TestTag.TWITTER_HANDLE_TEXT),
-                bodyText = speaker.twitterHandle,
-                textColor = MaterialTheme.chaiColorsPalette.secondaryButtonColor,
-            )
-        }
-    }
-}
-
-@Composable
-private fun SessionBannerImage(sessionDetails: SessionDetailsPresentationModel) {
-    AsyncImage(
-        model = sessionDetails.sessionImageUrl,
-        contentDescription = null,
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .wrapContentHeight()
-                .border(1.dp, ChaiTeal90, RoundedCornerShape(10.dp))
-                .testTag(TestTag.IMAGE_BANNER)
-                .clip(RoundedCornerShape(10.dp)),
-    )
-}
-
-@Composable
-private fun SessionSpeakerNameAndFavouriteIcon(
-    sessionDetails: SessionDetailsPresentationModel,
-    bookmarkSession: (String) -> Unit,
-    unBookmarkSession: (String) -> Unit,
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            imageVector = Icons.Filled.Android,
-            contentDescription = null,
-            modifier =
-                Modifier
-                    .height(14.dp)
-                    .width(15.dp),
-            tint = ChaiRed,
-        )
-
-        Spacer(modifier = Modifier.width(6.dp))
-
-        ChaiTextLabelLarge(
-            bodyText = stringResource(id = R.string.speaker_label),
-            textColor = ChaiRed,
-        )
-    }
-
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween,
-    ) {
-        ChaiTitle(
-            modifier = Modifier.testTag(TestTag.SPEAKER_NAME),
-            titleText = sessionDetails.speakers.joinToString(" & ") { it.name },
-            titleColor = MaterialTheme.chaiColorsPalette.textTitlePrimaryColor,
-        )
-
-        IconButton(
-            modifier = Modifier.size(32.dp),
-            onClick = {
-                if (sessionDetails.isStarred) {
-                    unBookmarkSession(sessionDetails.id)
-                } else {
-                    bookmarkSession(sessionDetails.id)
-                }
-            },
-        ) {
-            Icon(
-                modifier =
-                    Modifier
-                        .testTag(TestTag.FAVOURITE_ICON),
-                imageVector = if (sessionDetails.isStarred) Icons.Rounded.Star else Icons.Rounded.StarOutline,
-                contentDescription = stringResource(R.string.star_session_icon_description),
-                tint = if (sessionDetails.isStarred) ChaiRed else MaterialTheme.chaiColorsPalette.secondaryButtonColor,
-            )
-        }
-    }
-}
-
-@Composable
-private fun SessionTitleAndDescription(
-    sessionDetails: SessionDetailsPresentationModel,
-) {
-    ChaiBodyLargeBold(
-        modifier = Modifier.testTag(TestTag.SESSION_TITLE),
-        bodyText = sessionDetails.title,
-        textColor = MaterialTheme.chaiColorsPalette.textNormalColor,
-    )
-
-    Spacer(modifier = Modifier.height(15.dp))
-
-    ChaiBodyMedium(
-        modifier = Modifier.testTag(TestTag.SESSION_DESCRIPTION),
-        bodyText = sessionDetails.description,
-        textColor = MaterialTheme.chaiColorsPalette.textWeakColor,
-    )
-}
-
-@Composable
-private fun SessionLevel(sessionLevel: String) {
-    ChaiBodySmall(
-        modifier =
-            Modifier
-                .background(
-                    color = MaterialTheme.chaiColorsPalette.badgeBackgroundColor,
-                    shape = RoundedCornerShape(5.dp),
-                )
-                .padding(vertical = 3.dp, horizontal = 9.dp)
-                .testTag(TestTag.LEVEL),
-        bodyText = "#$sessionLevel".uppercase(),
-        textColor = ChaiWhite,
-    )
-}
-
-@Composable
-private fun SessionTimeAndRoom(
-    sessionDetails: SessionDetailsPresentationModel,
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        ChaiBodyXSmall(
-            modifier = Modifier.testTag(TestTag.TIME_SLOT),
-            bodyText = sessionDetails.timeSlot.uppercase(),
-            textColor = MaterialTheme.chaiColorsPalette.textWeakColor,
-        )
-        Spacer(modifier = Modifier.width(16.dp))
-        ChaiBodyXSmall(
-            bodyText = "|",
-            textColor = MaterialTheme.chaiColorsPalette.textWeakColor,
-        )
-        Spacer(modifier = Modifier.width(16.dp))
-        ChaiBodyXSmall(
-            modifier = Modifier.testTag(TestTag.ROOM),
-            bodyText = sessionDetails.venue.uppercase(),
-            textColor = MaterialTheme.chaiColorsPalette.textWeakColor,
-        )
-    }
-}
-
-@Composable
-private fun CustomDivider() {
-    Divider(
-        thickness = 1.dp,
-        color = MaterialTheme.chaiColorsPalette.surfaces,
-    )
-}
-
-@Composable
-private fun TopBar(onNavigationIconClick: () -> Unit) {
-    TopAppBar(
-        modifier = Modifier.testTag(TestTag.TOP_BAR),
-        colors =
-            TopAppBarDefaults.topAppBarColors(
-                containerColor = MaterialTheme.chaiColorsPalette.background,
-                navigationIconContentColor = MaterialTheme.chaiColorsPalette.textBoldColor,
-                scrolledContainerColor = MaterialTheme.chaiColorsPalette.background,
-                titleContentColor = MaterialTheme.chaiColorsPalette.textBoldColor,
-                actionIconContentColor = MaterialTheme.chaiColorsPalette.textBoldColor,
-            ),
-        title = {
-            ChaiBodyLarge(
-                bodyText = stringResource(id = R.string.session_details_label),
-                textColor = MaterialTheme.chaiColorsPalette.textBoldColor,
-            )
-        },
-        navigationIcon = {
-            IconButton(
-                onClick = { onNavigationIconClick() },
-            ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_back_arrow),
-                    contentDescription = stringResource(R.string.back_arrow_icon_description),
-                    tint = MaterialTheme.chaiColorsPalette.textBoldColor,
-                )
-            }
-        },
-    )
-}
-
-object TestTag {
-    private const val PREFIX = "sessionDetailsScreen:"
-
-    const val TOP_BAR = "$PREFIX topBar"
-    const val FLOATING_ACTION_BUTTON = "$PREFIX floatingActionButton"
-    const val SPEAKER_NAME = "$PREFIX speakerName"
-    const val FAVOURITE_ICON = "$PREFIX favouriteIcon"
-    const val SESSION_TITLE = "$PREFIX sessionTitle"
-    const val SESSION_DESCRIPTION = "$PREFIX sessionDescription"
-    const val IMAGE_BANNER = "$PREFIX imageBanner"
-    const val TIME_SLOT = "$PREFIX timeSlot"
-    const val ROOM = "$PREFIX room"
-    const val LEVEL = "$PREFIX level"
-    const val TWITTER_HANDLE_TEXT = "$PREFIX twitterHandleText"
 }
 
 @Preview(
