+++ presentation/src/main/java/com/android254/presentation/sessionDetails/view/components/TestTag.kt
@@ -1,0 +1,17 @@
+package com.android254.presentation.sessionDetails.view.components
+
+object TestTag {
+    private const val PREFIX = "sessionDetailsScreen:"
+
+    const val TOP_BAR = "$PREFIX topBar"
+    const val FLOATING_ACTION_BUTTON = "$PREFIX floatingActionButton"
+    const val SPEAKER_NAME = "$PREFIX speakerName"
+    const val FAVOURITE_ICON = "$PREFIX favouriteIcon"
+    const val SESSION_TITLE = "$PREFIX sessionTitle"
+    const val SESSION_DESCRIPTION = "$PREFIX sessionDescription"
+    const val IMAGE_BANNER = "$PREFIX imageBanner"
+    const val TIME_SLOT = "$PREFIX timeSlot"
+    const val ROOM = "$PREFIX room"
+    const val LEVEL = "$PREFIX level"
+    const val TWITTER_HANDLE_TEXT = "$PREFIX twitterHandleText"
+}
